requirements.txt
click
pydantic
structlog
tenacity
pyyaml
